Explicitly set game variables type value to int16_t

+++ src/vars/intvar.hpp
@@ -25,17 +25,17 @@
 
 struct intvar
 {
-    intvar(std::string const& p_name, int p_val) : name(p_name), val(p_val) { }
+    intvar(std::string const& p_name, int16_t p_val) : name(p_name), val(p_val) { }
 
     std::string name;
-    int val;
+    int16_t val;
 };
 
 namespace intvarm
 {
     void add_var_to_arr(std::vector<intvar>& p_vec, intvar& p_var);
-    int return_value(std::string const& p_name, std::vector<intvar>& p_vec);
-    void set_value(int p_val, std::string const& p_name, std::vector<intvar>& p_vec);
+    int16_t return_value(std::string const& p_name, std::vector<intvar>& p_vec);
+    void set_value(int16_t p_val, std::string const& p_name, std::vector<intvar>& p_vec);
 }
 
 #endif
